Updated Individual Project Proposal
Specified the broad and specific question that I would answer in this proposal. I also did a summary of the datasets, explaining the observations and variables of each and how this data will help me answer the questions that I selected.

I also wrangled the data to convert timestamps to figure out the duration. These basic wrangling techniques helped me better understand and interpret the data.